Showcase some challenges around making data-persist feel good
I tried to illustrate a common use case with a side menu, where when
you click on a button (usually a hamburger menu),
the menu collapses or expands.

The issue now is that in the Datastar version, we get a frame where
the navbar is rendered in the wrong state. We could, of course,
render the signal with the correct state and persist it on the server.

The problem is that it becomes a bit more work than it should be,
the initial thought might be to simply persist the state per user.

But if you have cases where a user is logged in on multiple devices
with different form factors, they’ll need to reconfigure
the stored state every time they load a new page on each device.
You could imagine that on a phone, you probably never want the side menu
to be expanded, while on a desktop, it might be useful to always have
it expanded.

Alternatively, we could store the state tied to the user’s session
and render with the correct signal from the first frame.
But that seems like a high price to pay for something that, in my opinion,
would be simplest to handle on the client.

Is there a way to get the best of both worlds — hydrate the signals
in the engine before we render the first node in the DOM?